+++ CHANGELOG.md
@@ -6,6 +6,32 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.1.0] 2021-03-02
+### Changed
+ * Fix Atomic Card entity & Formatter (add missing keywords & remove uuid)
+ * Some code style / PSR fix
+ * Add typehint on some files
+
+## [1.0.0] 2021-03-02
+### Changed
+ * now require php 7.4
+ * Update composer.json
+ * Fix tests
+
+----
+
+## [0.3.0] 2020-10-30
+### Changed
+ * Upgrade dependecy curl & require php 7.4
+
+## [0.2.0] 2020-09-04 - Release v0.2.0
+### Changed
+ * Regroup all client in one (MtgJsonClient)
+ * Update tests
+ * Update examples
+ * Update code
+
+## [0.1.0] 2020-09-04 - Release v0.1.0
 ### Added
  * Client for Set
  * Formatter